+++ server/src/app/blogs/add-blog/page.tsx
@@ -1,6 +1,15 @@
+import { redirect } from 'next/navigation'
+
+import { auth } from '@/auth'
 import { Blog, DashboardHeader } from '@/components/shared'
 
-const AddBlogPage = () => {
+const AddBlogPage = async () => {
+	const session = await auth()
+
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
+
 	return (
 		<div className='add-contents-page'>
 			<DashboardHeader title='Add' subtitle='Blog' breadcrumbs={['blogs', 'add-blog']} />
