+++ server/src/app/blogs/edit/[...id]/page.tsx
@@ -1,35 +1,15 @@
-'use client'
-
-import axios from 'axios'
 import Head from 'next/head'
-import { useParams } from 'next/navigation'
-import { useState, useEffect } from 'react'
+import { redirect } from 'next/navigation'
 
-import { IBlog } from '@/models/blog'
-import { Blog, DashboardHeader } from '@/components/shared'
+import { auth } from '@/auth'
+import { EditBlogsView } from './_components/edit-blogs-view'
 
-const EditBlogPage = () => {
-	const { id } = useParams() as { id: string }
+const EditBlogPage = async () => {
+	const session = await auth()
 
-	const [productInfo, setProductInfo] = useState<IBlog | null>(null)
-
-	useEffect(() => {
-		if (!id) {
-			return
-		}
-
-		const fetchProduct = async () => {
-			try {
-				const res = await axios.get(`/api/blogs?id=${id}`)
-
-				setProductInfo(res.data)
-			} catch (error) {
-				console.error('[BLOGS_EDIT] Data boot error:', error)
-			}
-		}
-
-		fetchProduct()
-	}, [id])
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
 
 	return (
 		<>
@@ -37,11 +17,7 @@ const EditBlogPage = () => {
 				<title>Update Blog</title>
 			</Head>
 
-			<div className='content-page'>
-				<DashboardHeader title='Edit' subtitle={productInfo?.title || ''} breadcrumbs={['blogs']} />
-
-				<div className='mt-12'>{productInfo && <Blog {...productInfo} />}</div>
-			</div>
+			<EditBlogsView />
 		</>
 	)
 }
