+++ server/src/app/gallery/add-photo/page.tsx
@@ -1,6 +1,15 @@
+import { redirect } from 'next/navigation'
+
+import { auth } from '@/auth'
 import { Photo, DashboardHeader } from '@/components/shared'
 
-const AddPhotoPage = () => {
+const AddPhotoPage = async () => {
+	const session = await auth()
+
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
+
 	return (
 		<div className='add-contents-page'>
 			<DashboardHeader title='Add' subtitle='Photo' breadcrumbs={['gallery', 'add-photo']} />
