+++ server/src/app/gallery/edit/[...id]/page.tsx
@@ -1,35 +1,15 @@
-'use client'
-
-import axios from 'axios'
 import Head from 'next/head'
-import { useParams } from 'next/navigation'
-import { useState, useEffect } from 'react'
+import { redirect } from 'next/navigation'
 
-import { IPhoto } from '@/models/photo'
-import { Photo, DashboardHeader } from '@/components/shared'
+import { auth } from '@/auth'
+import { EditPhotoView } from './_components/edit-photo-view'
 
-const EditPhotoPage = () => {
-	const { id } = useParams() as { id: string }
+const EditPhotoPage = async () => {
+	const session = await auth()
 
-	const [productInfo, setProductInfo] = useState<IPhoto | null>(null)
-
-	useEffect(() => {
-		if (!id) {
-			return
-		}
-
-		const fetchProduct = async () => {
-			try {
-				const res = await axios.get(`/api/photos?id=${id}`)
-
-				setProductInfo(res.data)
-			} catch (error) {
-				console.error('[PHOTOS_EDIT] Data boot error:', error)
-			}
-		}
-
-		fetchProduct()
-	}, [id])
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
 
 	return (
 		<>
@@ -37,11 +17,7 @@ const EditPhotoPage = () => {
 				<title>Update Photo</title>
 			</Head>
 
-			<div className='content-page'>
-				<DashboardHeader title='Edit' subtitle={productInfo?.title || ''} breadcrumbs={['gallery']} />
-
-				<div className='mt-12'>{productInfo && <Photo {...productInfo} />}</div>
-			</div>
+			<EditPhotoView />
 		</>
 	)
 }
