+++ server/src/app/projects/add-project/page.tsx
@@ -1,6 +1,15 @@
+import { redirect } from 'next/navigation'
+
+import { auth } from '@/auth'
 import { Project, DashboardHeader } from '@/components/shared'
 
-const AddProjectPage = () => {
+const AddProjectPage = async () => {
+	const session = await auth()
+
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
+
 	return (
 		<div className='add-contents-page'>
 			<DashboardHeader title='Add' subtitle='Project' breadcrumbs={['projects', 'add-project']} />
