+++ server/src/app/projects/edit/[...id]/page.tsx
@@ -1,35 +1,15 @@
-'use client'
-
-import axios from 'axios'
 import Head from 'next/head'
-import { useParams } from 'next/navigation'
-import { useState, useEffect } from 'react'
+import { redirect } from 'next/navigation'
 
-import { IProject } from '@/models/project'
-import { Project, DashboardHeader } from '@/components/shared'
+import { auth } from '@/auth'
+import { EditProjectView } from './_components/edit-project-view'
 
-const EditProjectPage = () => {
-	const { id } = useParams() as { id: string }
+const EditProjectPage = async () => {
+	const session = await auth()
 
-	const [productInfo, setProductInfo] = useState<IProject | null>(null)
-
-	useEffect(() => {
-		if (!id) {
-			return
-		}
-
-		const fetchProduct = async () => {
-			try {
-				const res = await axios.get(`/api/projects?id=${id}`)
-
-				setProductInfo(res.data)
-			} catch (error) {
-				console.error('[PROJECTS_EDIT] Data boot error:', error)
-			}
-		}
-
-		fetchProduct()
-	}, [id])
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
 
 	return (
 		<>
@@ -37,11 +17,7 @@ const EditProjectPage = () => {
 				<title>Update Project</title>
 			</Head>
 
-			<div className='content-page'>
-				<DashboardHeader title='Edit' subtitle={productInfo?.title || ''} breadcrumbs={['projects']} />
-
-				<div className='mt-12'>{productInfo && <Project {...productInfo} />}</div>
-			</div>
+			<EditProjectView />
 		</>
 	)
 }
