+++ server/src/app/shop/edit/[...id]/page.tsx
@@ -1,35 +1,15 @@
-'use client'
-
-import axios from 'axios'
 import Head from 'next/head'
-import { useParams } from 'next/navigation'
-import { useState, useEffect } from 'react'
+import { redirect } from 'next/navigation'
 
-import { IShop } from '@/models/shop'
-import { Shop, DashboardHeader } from '@/components/shared'
+import { auth } from '@/auth'
+import { EditProductView } from './_components/edit-product-view'
 
-const EditProductPage = () => {
-	const { id } = useParams() as { id: string }
+const EditProductPage = async () => {
+	const session = await auth()
 
-	const [productInfo, setProductInfo] = useState<IShop | null>(null)
-
-	useEffect(() => {
-		if (!id) {
-			return
-		}
-
-		const fetchProduct = async () => {
-			try {
-				const res = await axios.get(`/api/shops?id=${id}`)
-
-				setProductInfo(res.data)
-			} catch (error) {
-				console.error('[PRODUCTS_EDIT] Data boot error:', error)
-			}
-		}
-
-		fetchProduct()
-	}, [id])
+	if (!session) {
+		redirect('/auth/sign-in')
+	}
 
 	return (
 		<>
@@ -37,11 +17,7 @@ const EditProductPage = () => {
 				<title>Update Product</title>
 			</Head>
 
-			<div className='content-page'>
-				<DashboardHeader title='Edit' subtitle={productInfo?.title || ''} breadcrumbs={['shop']} />
-
-				<div className='mt-12'>{productInfo && <Shop {...productInfo} />}</div>
-			</div>
+			<EditProductView />
 		</>
 	)
 }
