+++ server/src/auth.ts
@@ -1,5 +1,19 @@
 import NextAuth from 'next-auth'
 
-import fullConfig from '@/auth.full.config'
+import authConfig from '@/auth.config'
 
-export const { auth, handlers, signIn, signOut } = NextAuth(fullConfig)
+export const { auth, handlers, signIn, signOut } = NextAuth({
+	secret: process.env.AUTH_SECRET,
+
+	session: {
+		strategy: 'jwt',
+		maxAge: 60 * 60 * 24 * 7, // 7 days
+		updateAge: 60 * 60 * 24, // 1 day
+	},
+
+	pages: {
+		signIn: '/auth/sign-in',
+	},
+
+	...authConfig,
+})
